requirements.txt
requests==2.32.3
cloudscraper==1.2.71
beautifulsoup4==4.12.3
python-dotenv==1.0.1
APScheduler==3.10.4
